requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.2
